+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- Columnar result returning from multiple blocks. Columns must be concatenated.
 
 ## [0.0.7] - 2017-10-12
 ### Added